feat: add CLI interface with new, open, search, import, sync commands
Unify CLI into main binary using clap derive macros. Commands:
- ztlgr new <path> - Create vault with Zettelkasten structure
- ztlgr open [path] - Open vault in TUI
- ztlgr search <query> - Full-text search via FTS5
- ztlgr import <source> - Import notes from directory
- ztlgr sync - Sync vault files with database
- Global flags: --vault, --format, --config, --verbose

No args falls back to interactive SetupWizard (backward compatible).

Also fixes 65+ pre-existing clippy warnings across the codebase.
Removes stub binaries (ztlgr-cli.rs, ztlgr.rs).

Tests: 264 passing (+16 new CLI tests)

+++ src/ui/widgets/modals/generic_modal.rs
@@ -84,8 +84,8 @@ impl GenericModal {
         let area = f.size();
 
         // Calculate modal dimensions (centered)
-        let modal_width = (area.width * self.width_percent as u16) / 100;
-        let modal_height = (area.height * self.height_percent as u16) / 100;
+        let modal_width = (area.width * self.width_percent) / 100;
+        let modal_height = (area.height * self.height_percent) / 100;
 
         let modal_x = (area.width.saturating_sub(modal_width)) / 2;
         let modal_y = (area.height.saturating_sub(modal_height)) / 2;
